requirements.txt
pandas>=2.1.0
holidays>=0.36
python-dateutil>=2.8.2
tabulate>=0.9.0
pytz>=2023.3
